requirements.txt
# Core dependencies
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# AI/ML Models
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Advanced RAG dependencies
rank-bm25>=0.2.2
transformers>=4.30.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.10.0

# Utilities
python-dotenv>=1.0.0
